manual_filtering/requirements.txt
Pillow>=9.0.0
send2trash>=1.8.0
pillow-heif>=0.10.0
